feat(workflow): edit and delete workflows real time p1
Part 1:
- Users on workflow while edit will be notified BUT not redirected (seen with @pguichon, need to find a trick to avoid user being redirected and lose all it's work).
- Users on workflow or in a story while deletion will be redirected to selected workflow if move stories was selected, otherwise to the first workflow.
- Delete and edit both update side-nav.

+++ src/app/workspace/project-detail/kanban-wrapper/project-kanban/project-kanban.component.ts
@@ -375,25 +375,30 @@ export class ProjectKanbanComponent {
 
   openDeleteWorkflowDialog() {
     const selectedWorkflow = this.workflowService.selectedEntity();
-    const dialogRef = this.dialog.open(DeleteWorkflowDialogComponent, {
-      ...matDialogConfig,
-      data: {
-        workflowName: selectedWorkflow.name,
-        workflowId: selectedWorkflow.id,
-      },
-    });
-    dialogRef.afterClosed().subscribe(async (formValue?: DeleteWorkflowFormData) => {
-      if (formValue) {
-        const moveToWorkflow: string | undefined = formValue.stories === "move" ? formValue.workflowTarget : undefined;
-        const deletedData = await this.projectKanbanService.deletesSelectedWorkflow(moveToWorkflow);
-        if (deletedData) {
-          await this.router.navigate(["..", deletedData.redirectionSlug], { relativeTo: this.activatedRoute });
-          this.notificationService.success({
-            title: "notification.workflow.deleted",
-            translocoTitleParams: { name: selectedWorkflow.name },
-          });
+    if (selectedWorkflow) {
+      const dialogRef = this.dialog.open(DeleteWorkflowDialogComponent, {
+        ...matDialogConfig,
+        data: {
+          workflowName: selectedWorkflow.name,
+          workflowId: selectedWorkflow.id,
+        },
+      });
+      dialogRef.afterClosed().subscribe(async (formValue?: DeleteWorkflowFormData) => {
+        if (formValue) {
+          const moveToWorkflow: string | undefined =
+            formValue.stories === "move" ? formValue.workflowTarget : undefined;
+          const deletedData = await this.projectKanbanService.deletesSelectedWorkflow(moveToWorkflow);
+          if (deletedData) {
+            await this.router.navigate(["..", deletedData.redirectionSlug], { relativeTo: this.activatedRoute });
+            this.notificationService.success({
+              title: "notification.workflow.deleted",
+              translocoTitleParams: { name: selectedWorkflow.name },
+            });
+          }
         }
-      }
-    });
+      });
+    } else {
+      throw new Error("[WORKFLOW] Cannot delete workflow, missing SelectedWorkflow");
+    }
   }
 }
